Polymorphic functions
Give type variables names 'a, 'b, etc.

In the unifier, make Variable implement Comparable, so that a is before b
is before z is before T0 is before T1 is before T10 etc. (Variables of
the form "T<integer>" are sorted after explicitly named variables, and
are compared using their integer ordinals.)